refactor: update service
