+++ pkg/api/service/entity.go
@@ -3,39 +3,20 @@ package service
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"net/url"
-	"strings"
 
 	"github.com/tkeel-io/core/pkg/entities"
 	"github.com/tkeel-io/core/pkg/service"
 
-	dapr "github.com/dapr/go-sdk/client"
 	"github.com/dapr/go-sdk/service/common"
 	"github.com/pkg/errors"
 )
-
-const (
-	entityFieldType   = "type"
-	entityFieldID     = "id"
-	entityFieldOwner  = "owner"
-	entityFieldSource = "source"
-)
-
-var errBodyMustBeJSON = errors.New("request body must be json")
-
-func entityFieldRequired(fieldName string) error {
-	return fmt.Errorf("entity field(%s) required", fieldName)
-}
-
-type Entity = entities.EntityBase
 
 // EntityService is a time-series service.
 type EntityService struct {
 	ctx           context.Context
 	cancel        context.CancelFunc
-	daprClient    dapr.Client
 	entityManager *entities.EntityManager
 }
 
@@ -43,17 +24,11 @@ type EntityService struct {
 func NewEntityService(ctx context.Context, mgr *entities.EntityManager) (*EntityService, error) {
 	ctx, cancel := context.WithCancel(ctx)
 
-	cli, err := dapr.NewClient()
-	if err != nil {
-		err = errors.Wrap(err, "start dapr clint err")
-	}
-
 	return &EntityService{
 		ctx:           ctx,
 		cancel:        cancel,
-		daprClient:    cli,
 		entityManager: mgr,
-	}, err
+	}, nil
 }
 
 // Name return the name.
@@ -70,78 +45,7 @@ func (e *EntityService) RegisterService(daprService common.Service) (err error) 
 	if err = daprService.AddServiceInvocationHandler("/plugins/{plugin}/entities", e.entitiesHandler); nil != err {
 		return
 	}
-	if err = e.AddSubTopic(daprService, "core-pub", "core-pubsub"); nil != err {
-		return
-	}
 	return
-}
-
-func (e *EntityService) AddSubTopic(daprService common.Service, topic, pubsubName string) (err error) {
-	sub := &common.Subscription{
-		PubsubName: pubsubName,
-		Topic:      topic,
-		Route:      "/",
-		Metadata:   map[string]string{},
-	}
-	if err = daprService.AddTopicEventHandler(sub, e.topicHandler); err != nil {
-		return
-	}
-	return
-}
-
-func getSourceFrom(pubsubName string) (source string) {
-	return strings.Split(pubsubName, "-")[0]
-}
-
-func TopicEvent2EntityContext(in *common.TopicEvent) (out *entities.EntityContext, err error) {
-	ec := entities.EntityContext{}
-	_ = getSourceFrom(in.PubsubName)
-	var entityID, owner string
-	ec.Headers = make(map[string]string)
-	if in.DataContentType == "application/json" {
-		inData, ok := in.Data.(map[string]interface{})
-		if !ok {
-			return nil, errTypeError
-		}
-		switch entityIds := inData["entity_id"].(type) {
-		case string:
-			entityID = entityIds
-		default:
-			return nil, errTypeError
-		}
-		switch tempOwner := inData["owner"].(type) {
-		case string:
-			owner = tempOwner
-		default:
-			err = errTypeError
-			return
-		}
-		switch tempData := inData["data"].(type) {
-		case string, []byte:
-			values := make(map[string]interface{})
-			values["__data__"] = tempData
-			ec.Message = &entities.EntityMessage{SourceID: entityID, Values: values}
-		case map[string]interface{}:
-			ec.Message = &entities.EntityMessage{SourceID: entityID, Values: tempData}
-		default:
-			err = errTypeError
-			return
-		}
-
-		ec.Headers["user_id"] = owner
-		ec.SetTarget(entityID)
-	}
-	return &ec, nil
-}
-
-func (e *EntityService) topicHandler(ctx context.Context, in *common.TopicEvent) (retry bool, err error) {
-	if ec, err := TopicEvent2EntityContext(in); err != nil {
-		return false, err
-	} else if in.DataContentType == "application/json" {
-		e.entityManager.SendMsg(*ec)
-	}
-
-	return false, nil
 }
 
 // Echo test for RegisterService.
@@ -156,15 +60,10 @@ func (e *EntityService) entityHandler(ctx context.Context, in *common.Invocation
 	switch in.Verb {
 	case http.MethodGet:
 		return e.entityGet(ctx, in)
-	case http.MethodPost:
-		return e.entityCreate(ctx, in)
 	case http.MethodPut:
 		return e.entityUpdate(ctx, in)
 	case http.MethodDelete:
 		return e.entityDelete(ctx, in)
-		// temporary.
-	case http.MethodPatch:
-		return e.entityCreate(ctx, in)
 	default:
 	}
 
@@ -321,7 +220,7 @@ func (e *EntityService) entityUpdate(ctx context.Context, in *common.InvocationE
 
 	defer errResult(out, err)
 
-	_, err = e.getEntityFrom(ctx, entity, in, false)
+	_, err = e.getEntityFrom(ctx, entity, in, true)
 	if nil != err {
 		return
 	}
@@ -357,7 +256,7 @@ func (e *EntityService) entityDelete(ctx context.Context, in *common.InvocationE
 
 	defer errResult(out, err)
 
-	_, err = e.getEntityFrom(ctx, entity, in, false)
+	_, err = e.getEntityFrom(ctx, entity, in, true)
 	if nil != err {
 		return
 	}
@@ -374,6 +273,19 @@ func (e *EntityService) entityDelete(ctx context.Context, in *common.InvocationE
 	return
 }
 
+// EntityList List entities.
+func (e *EntityService) entityList(ctx context.Context, in *common.InvocationEvent) (out *common.Content, err error) {
+	out = &common.Content{
+		Data:        in.Data,
+		ContentType: in.ContentType,
+		DataTypeURL: in.DataTypeURL,
+	}
+	// TODO
+
+	defer errResult(out, err)
+	return
+}
+
 // Echo test for RegisterService.
 func (e *EntityService) entitiesHandler(ctx context.Context, in *common.InvocationEvent) (out *common.Content, err error) {
 	if in == nil {
@@ -384,13 +296,8 @@ func (e *EntityService) entitiesHandler(ctx context.Context, in *common.Invocati
 	switch in.Verb {
 	case http.MethodPost:
 		return e.entityCreate(ctx, in)
-	case http.MethodPut:
-		return e.entityUpdate(ctx, in)
-	case http.MethodDelete:
-		return e.entityDelete(ctx, in)
-		// temporary.
-	case http.MethodPatch:
-		return e.entityCreate(ctx, in)
+	case http.MethodGet:
+		return e.entityList(ctx, in)
 	default:
 	}
