+++ pkg/bootstrap/server.go
@@ -100,6 +100,7 @@ func initAPIRegistry(ctx context.Context, apiRegistry *api.Registry, apiConfig *
 		err       error
 		eventAPI  *service.EventService
 		entityAPI *service.EntityService
+		subAPI    *service.SubService
 	)
 
 	// register event api.
@@ -130,5 +131,12 @@ func initAPIRegistry(ctx context.Context, apiRegistry *api.Registry, apiConfig *
 		return errors.Wrap(err, "api registry add service err")
 	}
 
+	if subAPI, err = service.NewSubService(ctx, entityManager); err != nil {
+		return errors.Wrap(err, "new sub service err")
+	}
+
+	if err = apiRegistry.AddService(subAPI); err != nil {
+		return errors.Wrap(err, "api registry add service err")
+	}
 	return nil
 }
